Initial commit
Basic outline of the extension.
