+++ settings.gradle
@@ -4,17 +4,17 @@ pluginManagement {
         // Specifying it here, rather than build.gradle, makes it possible
         // to include the extension as a subproject of QuPath itself
         // (which is useful during development)
-        id 'org.bytedeco.gradle-javacpp-platform' version '1.5.9'
+        id 'org.bytedeco.gradle-javacpp-platform' version '1.5.10'
     }
 }
 
 // TODO: Name your QuPath extension here!
-rootProject.name = 'my-qupath-extension'
+rootProject.name = 'qupath-extension-training'
 
 // TODO: Define the QuPath version compatible with the extension
 // Note that the QuPath API isn't stable; something designed for
 // 0.X.a should work with 0.X.b, but not necessarily with 0.Y.a.
-gradle.ext.qupathVersion = "0.5.0"
+gradle.ext.qupathVersion = "0.6.0-SNAPSHOT"
 
 dependencyResolutionManagement {
