--- src/main/resources/qupath/ext/template/ui/interface.fxml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<fx:root type="VBox" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" xmlns="http://javafx.com/javafx/20" xmlns:fx="http://javafx.com/fxml/1">
-    <Button onAction="#runDemoExtension" text="Run"/>
-    <Spinner fx:id="threadSpinner" prefWidth="75.0">
-        <valueFactory>
-            <SpinnerValueFactory.IntegerSpinnerValueFactory max="96" min="1" />
-        </valueFactory>
-    </Spinner>
-</fx:root>
