--- src/main/resources/qupath/ext/template/ui/strings.properties
@@ -1,3 +1,0 @@
-title = Demo extension
-
-threads = Threads set to %d
