+++ src/main/resources/qupath/ext/training/ui/strings.properties
@@ -1,0 +1,2 @@
+ext.title = QuPath training extension
+ext.description = An extension to help learn how to use QuPath
